# docs/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "MAIN Smart Parking",
    "version": "1.0.0",
    "description": "SRS 2.3. ADMIN includes USER privileges; OWNER means resource ownership or ADMIN. Business ALLOW/DENY both return 200. Timestamps UTC; history ranges [from,to)."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
    "/auth/csrf": {
      "get": {
        "operationId": "get__auth_csrf",
        "tags": [
          "PUBLIC"
        ],
        "summary": "GET /auth/csrf",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsrfView"
                }
              }
            }
          },
          "400": {
            "description": "Validation / signature error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Role / CSRF denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Missing or not owned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Business or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Registration/login rate limit retry delay, seconds."
              }
            }
          },
          "503": {
            "description": "Retryable infrastructure failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": []
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "post__auth_register",
        "tags": [
          "PUBLIC"
        ],
        "summary": "POST /auth/register",
        "security": [],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserView"
                }
              }
            }
          },
          "400": {
            "description": "Validation / signature error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Role / CSRF denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Missing or not owned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Business or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Registration/login rate limit retry delay, seconds."
              }
            }
          },
          "503": {
            "description": "Retryable infrastructure failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-TOKEN",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Obtain fresh context after login/logout via /auth/csrf."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "post__auth_login",
        "tags": [
          "PUBLIC"
        ],
        "summary": "POST /auth/login",
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserView"
                }
              }
            }
          },
          "400": {
            "description": "Validation / signature error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Role / CSRF denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Missing or not owned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Business or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Registration/login rate limit retry delay, seconds."
              }
            }
          },
          "503": {
            "description": "Retryable infrastructure failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-TOKEN",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Obtain fresh context after login/logout via /auth/csrf."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "post__auth_logout",
        "tags": [
          "USER"
        ],
        "summary": "POST /auth/logout",
        "security": [
          {
            "Session": []
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Validation / signature error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Role / CSRF denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Missing or not owned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Business or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Registration/login rate limit retry delay, seconds."
              }
            }
          },
          "503": {
            "description": "Retryable infrastructure failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-TOKEN",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Obtain fresh context after login/logout via /auth/csrf."
          }
        ]
      }
    },
    "/me": {
      "get": {
        "operationId": "get__me",
        "tags": [
          "USER"
        ],
        "summary": "GET /me",
        "security": [
          {
            "Session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserView"
                }
              }
            }
          },
          "400": {
            "description": "Validation / signature error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Role / CSRF denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Missing or not owned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Business or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Registration/login rate limit retry delay, seconds."
              }
            }
          },
          "503": {
            "description": "Retryable infrastructure failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": []
      },
      "patch": {
        "operationId": "patch__me",
        "tags": [
          "USER"
        ],
        "summary": "PATCH /me",
        "security": [
          {
            "Session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserView"
                }
              }
            }
          },
          "400": {
            "description": "Validation / signature error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Role / CSRF denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Missing or not owned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Business or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Registration/login rate limit retry delay, seconds."
              }
            }
          },
          "503": {
            "description": "Retryable infrastructure failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-TOKEN",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Obtain fresh context after login/logout via /auth/csrf."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileRequest"
              }
            }
          }
        }
      }
    },
    "/me/password": {
      "put": {
        "operationId": "put__me_password",
        "tags": [
          "USER"
        ],
        "summary": "PUT /me/password",
        "security": [
          {
            "Session": []
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Validation / signature error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Role / CSRF denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Missing or not owned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Business or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Registration/login rate limit retry delay, seconds."
              }
            }
          },
          "503": {
            "description": "Retryable infrastructure failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-TOKEN",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Obtain fresh context after login/logout via /auth/csrf."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "FieldError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "fieldErrors",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "ALREADY_AUTHENTICATED",
              "ALREADY_INSIDE",
              "AUTH_REQUIRED",
              "BODY_TOO_LARGE",
              "CARD_BLOCKED",
              "CSRF_INVALID",
              "CURRENT_PASSWORD_INVALID",
              "DATABASE_UNAVAILABLE",
              "DEVICE_BOOT_NOT_REGISTERED",
              "DEVICE_DISABLED",
              "DEVICE_RATE_LIMITED",
              "DEVICE_UNAUTHORIZED",
              "ENABLED_CARD_EXISTS",
              "FORBIDDEN",
              "IDEMPOTENCY_CONFLICT",
              "INVALID_CREDENTIALS",
              "INVALID_SESSION_STATE",
              "INVALID_SIGNATURE",
              "LOGIN_RATE_LIMITED",
              "NOT_FOUND",
              "ORDER_IN_PROGRESS",
              "PACKAGE_DISABLED",
              "PAYMENT_CONSISTENCY_ERROR",
              "PAYMENT_MODE_MISMATCH",
              "PAYMENT_PROVIDER_UNAVAILABLE",
              "PROVIDER_MISMATCH",
              "RECONCILE_RATE_LIMITED",
              "REGISTRATION_RATE_LIMITED",
              "SESSION_CAPACITY_CONFLICT",
              "SESSION_INVALID",
              "UID_TAKEN",
              "USERNAME_TAKEN",
              "USER_DISABLED",
              "VALIDATION_ERROR"
            ]
          },
          "message": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "traceId": {
            "type": "string"
          },
          "orderId": {
            "type": "string",
            "format": "uuid",
            "description": "Existing order for ORDER_IN_PROGRESS; only returned within authorized ownership."
          }
        }
      },
      "UserView": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "username",
          "fullName",
          "phone",
          "role",
          "status",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DISABLED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CsrfView": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "token",
          "headerName"
        ],
        "properties": {
          "token": {
            "type": "string"
          },
          "headerName": {
            "type": "string"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "username",
          "password",
          "fullName"
        ],
        "properties": {
          "username": {
            "type": "string",
            "minLength": 3,
            "maxLength": 50,
            "description": "Trim and lowercase, then ASCII [a-z0-9._-]."
          },
          "password": {
            "type": "string",
            "writeOnly": true,
            "description": "10–72 UTF-8 bytes for a new password; never trim."
          },
          "fullName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "maxLength": 20
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "writeOnly": true,
            "description": "10–72 UTF-8 bytes for a new password; never trim."
          }
        }
      },
      "ProfileRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "fullName"
        ],
        "properties": {
          "fullName": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "PasswordRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "properties": {
          "currentPassword": {
            "type": "string",
            "writeOnly": true,
            "description": "10–72 UTF-8 bytes for a new password; never trim."
          },
          "newPassword": {
            "type": "string",
            "writeOnly": true,
            "description": "10–72 UTF-8 bytes for a new password; never trim."
          }
        }
      },
...
    },
    "securitySchemes": {
      "Session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID"
      },
      "DeviceId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Device-Id"
      },
      "DeviceKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Device-Key"
      }
    }
  }
}